feat: inicio Inertia

+++ database/migrations/2023_04_10_171922_create_recetas_ingredientes_table.php
@@ -15,11 +15,11 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('receta_id');
             $table->unsignedBigInteger('ingrediente_id');
-            $table->integer('cantidad')->nullable();
-            $table->integer('kcal')->nullable();
-            $table->integer('proteinas')->nullable();
-            $table->integer('grasas')->nullable();
-            $table->integer('carbohidratos')->nullable();
+            $table->double('cantidad')->nullable();
+            $table->double('kcal')->nullable();
+            $table->double('proteinas')->nullable();
+            $table->double('grasas')->nullable();
+            $table->double('carbohidratos')->nullable();
             $table->timestamps();
             $table->foreign('receta_id')->references('id')->on('recetas')->onDelete('cascade');
             $table->foreign('ingrediente_id')->references('id')->on('ingredientes')->onDelete('cascade');
